fix] to keep the client compatible with qt4.5
git-svn-id: svn+ssh://rcs.lan-puteaux.avencall.com/var/lib/svn/xivo/trunk/cticlients@8060 dc85baef-7018-0410-bfbd-8a639da648bb

+++ common/queuespanel.cpp
@@ -270,7 +270,7 @@ void QueuesPanel::eatQueuesStats(QVariantMap p)
 
             if (duration_stats.contains(stats)) {
                 if (p["stats"].toMap()[queueid].toMap()[stats].toString() != "na") {
-                    int sec_total = qRound(p["stats"].toMap()[queueid].toMap()[stats].toFloat());
+                    int sec_total = qRound(p["stats"].toMap()[queueid].toMap()[stats].toDouble());
                     __format_duration(&field, sec_total);
                 } else {
                     field = p["stats"].toMap()[queueid].toMap()[stats].toString();
